Taskes care of Issue #1 (https://goo.gl/E2g00L): Added implementations
