+++ src/main/java/constants/BoardConstants.java
@@ -1,0 +1,8 @@
+package constants;
+
+/**
+ * @author rachit
+ **/
+public class BoardConstants {
+    public static final Integer EDGE_LENGTH = 8;
+}
